Some small changes/fixes
Reverted Cat's changes to ammunition in 'AtkParam_Pc' and instead changed the 'Attack Element Correct ID' from bows, greatbows and crossbows in EquipParamWeapon.

Reduced the stability of the Leather Shield from 75 to 62 (the same as the rest of small shields).

Added air juggle stagger to the Dragonhead Greatshield WA.

Increased the stats of some character NPC so that they can meet the stat requirements of the equipment that they use.

Changed the summary of Frozen Chaos Essence and Odessa's Crystal so that they mention their effects instead of reinforcement.

Fixed some of the affixing prompts mentioning 'ring' and not 'accessory'.

Fixed the Lucky Serpentine Twindaggers and the enchanted/holy Moonlight Sword having the wrong damage listed in the Mod Data excel file (also removed the Lethal, Sharp and Blunt infusions from the Moonlight Sword).